requirements.txt
python-dotenv
pydantic
langchain-community
langchain-openai
langchain-anthropic
wikipedia
langchain
duckduckgo-search